requirements.txt
streamlit-folium>=0.27.0
folium>=0.17.0
pystac-client>=0.9.0
rasterio>=1.3.0
numpy>=1.24.0
matplotlib>=3.7.0
Pillow>=10.0.0
